+++ familyhelper-ui-view/packages/webapp-pc/src/views/nodes/elementPlus/hyperscript/placeholder/index.ts
@@ -1,7 +1,7 @@
 import { type Component, defineComponent, h } from 'vue'
 
 import PlaceholderPanel from '@/components/comvisual/layout/placeholderPanel/PlaceholderPanel.vue'
-import BorderLayoutPanel from '@/components/elementPlus/layout/borderLayoutPanel/BorderLayoutPanel.vue'
+import RootBorderLayoutPanel from '@/components/elementPlus/layout/rootBorderLayoutPanel/RootBorderLayoutPanel.vue'
 
 import styles from './index.module.css'
 
@@ -12,8 +12,11 @@ export function placeholder(placeholder: string): Component {
       return () =>
         h('div', { class: styles['placeholder-container'] }, [
           h(
-            BorderLayoutPanel,
-            {},
+            RootBorderLayoutPanel,
+            {
+              initialDockStatus: 4,
+              initialFloatyY: -200,
+            },
             {
               default: () => h(PlaceholderPanel, { text: placeholder }),
             },